rbd: fix fail to resize operation with integrity mode enabled
Problem: Resize is not possible if integrity mode is enabled.
Before a resize on a integrity protected encrypted volume
would only give the cryptsetup error message which is not very clear.

Solution: Check if integrity option is used in a cryptsetup status
before a cryptsetup resize to check if the volume is integrity
protected. If it is we fail the resize operation and throw a
semantically useful error message.

Signed-off-by: David Mohren <david.mohren@clyso.com>