nvmeof: fix already connected bug
when ConnectSubsystem() is called to connect to an already
connected subsystem listener <--> host, the ExecCommandWithTimeout returns
err "already connected". Instead of parsing the error message,
it is cleaner and more convenient to check the
connection state using isSubsystemConnected().
We always call ConnectSubsystem() with HostNQN,
so calling isPathConnected() without hostnqn does not make sense.
Therefore, the check is placed under if req.HostNQN.

Signed-off-by: gadi-didi <gadi.didi@ibm.com>